# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/k8s/deployment/query": {
            "get": {
                "tags": [
                    "K8S"
                ],
                "summary": "通过标签查询Deployment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "标签选择器，如：app.kubernetes.io/name=abtesting-ms",
                        "name": "label_selector",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "环境",
                        "name": "env",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/k8s.ApplicationStatus"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/k8s/pod/list": {
            "get": {
                "tags": [
                    "K8S"
                ],
                "summary": "获取命名空间中的所有pod列表（调试用）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "环境",
                        "name": "env",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/k8s.PodInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/k8s/pod/query": {
            "get": {
                "tags": [
                    "K8S"
                ],
                "summary": "获取应用的pod列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "应用名称",
                        "name": "app_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "环境",
                        "name": "env",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/k8s.PodInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "k8s.ApplicationStatus": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string"
                },
                "available_replicas": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "env": {
                    "type": "string"
                },
                "exists": {
                    "type": "boolean"
                },
                "image": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "ready_replicas": {
                    "type": "integer"
                },
                "replicas": {
                    "type": "integer"
                },
                "service_exists": {
                    "type": "boolean"
                },
                "service_port": {
                    "type": "integer"
                },
                "service_type": {
                    "type": "string"
                }
            }
        },
        "k8s.PodInfo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "host_ip": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "node_name": {
                    "type": "string"
                },
                "phase": {
                    "description": "Pod阶段：Pending、Running、Succeeded、Failed、Unknown",
                    "type": "string"
                },
                "pod_ip": {
                    "type": "string"
                },
                "restart_count": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "util.ResponseTemplate": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "此处约定：1代表成功，0代表失败",
                    "type": "integer"
                },
                "error": {
                    "description": "如果请求失败，这里一定要给出错误的信息"
                },
                "help": {
                    "description": "显示接口文档地址，便于别人排错",
                    "type": "string"
                },
                "message": {
                    "description": "对请求结果的描述消息，可以为空",
                    "type": "string"
                },
                "result": {
                    "description": "如果请求成功，这里给出成功的结果"
                }
            }
        },
...
    }
}